# perf(geometry): stop using numpy array in "uniform_frechet_distance"