Modifications:
1. Stop using the numpy array.
2. Use `math.dist` to calculate the distance of two vector.
3. Use `tuple` instead of `list` when possible.
4. Use more generator and iterator.
5. Simply the calculation logic.

Target: enhance the performance of the following methods:
- `Polyline2D.uniform_frechet_distance`
- `Polyline2D.similarity`

Results:
```python
polyline_1 = [[1, 1], [1, 2], [2, 2]]
polyline_2 = [[4, 5], [2, 1], [3, 3]]
```

- before:
```python
>>> %timeit Polyline2D.uniform_frechet_distance(polyline_1, polyline_2)
231 µs ± 2.77 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

>>> %timeit Polyline2D.similarity(polyline_1, polyline_2)
239 µs ± 3.36 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

- after:
```python
>>> %timeit Polyline2D.uniform_frechet_distance(polyline_1, polyline_2)
59.1 µs ± 859 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

>>> %timeit Polyline2D.similarity(polyline_1, polyline_2)
61 µs ± 380 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```